Update to v106r130 release.
byuu says:

I completed the initial emulation of the Famicom Disk System. Reading and
writing to disks work, plus the FM synthesis audio channel is now supported.
It's going to be far from perfect, but ... it mostly works.

A horrifying amount of additional work will be needed on the GUI to be able to
handle save states for FDS games cleanly.